# backend/docs/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Housing Society Management API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/amenities": {
      "get": {
        "tags": [
          "amenities"
        ],
        "summary": "List Amenities",
        "operationId": "list_amenities_api_v1_amenities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AmenityOut"
                  },
                  "type": "array",
                  "title": "Response List Amenities Api V1 Amenities Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "amenities"
        ],
        "summary": "Create Amenity",
        "operationId": "create_amenity_api_v1_amenities_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AmenityCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AmenityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/amenities/bookings": {
      "post": {
        "tags": [
          "amenities"
        ],
        "summary": "Create Booking",
        "operationId": "create_booking_api_v1_amenities_bookings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AmenityBookingCreateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AmenityBookingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "amenities"
        ],
        "summary": "List Bookings",
        "operationId": "list_bookings_api_v1_amenities_bookings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AmenityBookingOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/amenities/bookings/{booking_id}/decision": {
      "post": {
        "tags": [
          "amenities"
        ],
        "summary": "Decide Booking",
        "description": "Admin (any scope) or Sub-admin (property's scope) \u2014 Section 20.",
        "operationId": "decide_booking_api_v1_amenities_bookings__booking_id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booking Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AmenityBookingDecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AmenityBookingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/account-entries": {
      "post": {
        "tags": [
          "account-entries"
        ],
        "summary": "Create",
        "description": "Section 13.1 \u2014 sirf Admin entry create kar sakta hai.",
        "operationId": "create_api_v1_account_entries_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountEntryCreateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountEntryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "account-entries"
        ],
        "summary": "List All",
        "description": "Section 13.1 \u2014 sabhi Residents ko visible, read-only.",
        "operationId": "list_all_api_v1_account_entries_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AccountEntryOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/account-entries/balance": {
      "get": {
        "tags": [
          "account-entries"
        ],
        "summary": "Balance",
        "operationId": "balance_api_v1_account_entries_balance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceSummaryOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/account-entries/{entry_id}": {
      "patch": {
        "tags": [
          "account-entries"
        ],
        "summary": "Edit",
        "description": "Section 13.1 \u2014 edit history preserved, is_edited flag drives the\n'Edited' badge in the UI.",
        "operationId": "edit_api_v1_account_entries__entry_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountEntryUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountEntryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountEntryCreateIn": {
        "properties": {
          "entry_type": {
            "$ref": "#/components/schemas/EntryType"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "entry_date": {
            "type": "string",
            "format": "date",
            "title": "Entry Date"
          }
        },
        "type": "object",
        "required": [
          "entry_type",
          "title",
          "amount",
          "entry_date"
        ],
        "title": "AccountEntryCreateIn",
        "description": "MANUAL entries only \u2014 Admin-only (Section 13.1)."
      },
      "AccountEntryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "society_id": {
            "type": "string",
            "format": "uuid",
            "title": "Society Id"
          },
          "entry_type": {
            "$ref": "#/components/schemas/EntryType"
          },
          "source": {
            "$ref": "#/components/schemas/EntrySource"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "entry_date": {
            "type": "string",
            "format": "date",
            "title": "Entry Date"
          },
          "is_edited": {
            "type": "boolean",
            "title": "Is Edited"
          },
          "last_edited_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Edited At"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "society_id",
          "entry_type",
          "source",
          "title",
          "description",
          "amount",
          "entry_date",
          "is_edited",
          "last_edited_at",
          "created_by",
          "created_at"
        ],
        "title": "AccountEntryOut"
      },
      "AccountEntryUpdateIn": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "entry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Date"
          }
        },
        "type": "object",
        "title": "AccountEntryUpdateIn"
      },
      "AmenityBookingCreateIn": {
        "properties": {
          "amenity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Amenity Id"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "booking_date": {
            "type": "string",
            "format": "date",
            "title": "Booking Date"
          },
          "start_time": {
            "type": "string",
            "format": "time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "time",
            "title": "End Time"
          }
        },
        "type": "object",
        "required": [
          "amenity_id",
          "property_id",
          "booking_date",
          "start_time",
          "end_time"
        ],
        "title": "AmenityBookingCreateIn"
      },
      "AmenityBookingDecisionIn": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          }
        },
        "type": "object",
        "required": [
          "approve"
        ],
        "title": "AmenityBookingDecisionIn"
      },
      "AmenityBookingOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "society_id": {
            "type": "string",
            "format": "uuid",
            "title": "Society Id"
          },
          "amenity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Amenity Id"
          },
          "property_id": {
            "type": "string",
            "format": "uuid",
            "title": "Property Id"
          },
          "resident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Resident Id"
          },
          "booking_date": {
            "type": "string",
            "format": "date",
            "title": "Booking Date"
          },
          "start_time": {
            "type": "string",
            "format": "time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "time",
            "title": "End Time"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "society_id",
          "amenity_id",
          "property_id",
          "resident_id",
          "booking_date",
          "start_time",
          "end_time",
          "status",
          "approved_by",
          "created_at"
        ],
        "title": "AmenityBookingOut"
      },
      "AmenityCreateIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "AmenityCreateIn"
      },
      "AmenityOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "society_id": {
            "type": "string",
            "format": "uuid",
            "title": "Society Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "society_id",
          "name",
          "description",
          "is_active"
        ],
        "title": "AmenityOut"
      },
...
      "BalanceSummaryOut": {
        "properties": {
          "total_income": {
            "type": "number",
            "title": "Total Income"
          },
          "total_expense": {
            "type": "number",
            "title": "Total Expense"
          },
          "balance": {
            "type": "number",
            "title": "Balance"
          }
        },
        "type": "object",
        "required": [
          "total_income",
          "total_expense",
          "balance"
        ],
        "title": "BalanceSummaryOut",
        "description": "Section 13.0 \u2014 2-column ledger: Income, Expense, Balance = Income - Expense."
      },
      "BookingStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "APPROVED",
          "REJECTED",
          "CANCELLED"
        ],
        "title": "BookingStatus"
      },
...
      "EntrySource": {
        "type": "string",
        "enum": [
          "MANUAL",
          "MAINTENANCE_PAYMENT",
          "EXPENSE_BILL",
          "ADJUSTMENT"
        ],
        "title": "EntrySource"
      },
      "EntryType": {
        "type": "string",
        "enum": [
          "INCOME",
          "EXPENSE"
        ],
        "title": "EntryType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_AccountEntryOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AccountEntryOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "Page[AccountEntryOut]"
      },
      "Page_AmenityBookingOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AmenityBookingOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "Page[AmenityBookingOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}